new pages build script

+++ buildpages.sh
@@ -2,7 +2,8 @@ rm -fr /tmp/gh-pages
 mkdir /tmp/gh-pages
 cp -r src/main/webapp/ /tmp/gh-pages
 cp -r src/test/webapp/ /tmp/gh-pages
-git checkout gh-pages
-rm -fr *
-cp -r /tmp/gh-pages .
+if git checkout gh-pages; then
+  rm -fr *
+  cp -r /tmp/gh-pages/ .
+fi
